Refactored AOC into a utils file
