+++ uqcsbot/bot.py
@@ -30,6 +30,7 @@ class UQCSBot(commands.Bot):
         self.ADMIN_ALERTS_CNAME = "admin-alerts"
         self.GENERAL_CNAME = "general"
         self.AOC_CNAME = "contests"
+        self.AOC_ROLE = "CPG"
         self.BOT_TIMEZONE = timezone("Australia/Brisbane")
 
         self.uqcs_server: discord.Guild
